[TASK-2468] [*] Extract MemList remap helper, add UnsafeList.AddRange(Span<T>)
