+++ MemoryAllocator/State/StateParts/Destroy/State/Components/Kill/KillCallbackComponent.cs
@@ -7,9 +7,9 @@ namespace Sapientia.MemoryAllocator.State
 		public void OnEntityKilled(WorldState worldState, Entity callbackReceiver);
 	}
 
-	// children - kill-дерево, копир тащит как own. parents - ремап с дропом. callbackTargets - обратный
-	// индекс, ремап как обычный список ссылок. killCallbacks - ProxyPtr-подписки, копируются через
-	// ISubscriberCopyable.Copy (payload в новую арену + ремап Entity-полей внутри).
+	// children - kill-дерево, копир тащит его как собственное. parents - перенастройка с дропом. callbackTargets -
+	// обратный индекс, перенастраивается как обычный список ссылок. killCallbacks - ProxyPtr-подписки, копируются
+	// через ISubscriberCopyable.Copy (payload в новую арену + перенастройка Entity-полей внутри).
 	[ManualCopy]
 	public partial struct KillCallbackComponent : IComponent
 	{
