+++ MemoryAllocator/_scripts.generated/InterfaceProxyGenerator/IActionProxy.generated.cs
@@ -7,7 +7,7 @@ namespace Sapientia.TypeIndexer
 {
 	public unsafe struct IActionProxy : IProxy
 	{
-		public static readonly ProxyId ProxyId = 13;
+		public static readonly ProxyId ProxyId = 12;
 		ProxyId IProxy.ProxyId
 		{
 			[System.Runtime.CompilerServices.MethodImplAttribute(256)]
