+++ MemoryAllocator/_scripts.generated/InterfaceProxyGenerator/IElementDestroyHandlerProxy.generated.cs
@@ -7,7 +7,7 @@ namespace Sapientia.TypeIndexer
 {
 	public unsafe struct IElementDestroyHandlerProxy : IProxy
 	{
-		public static readonly ProxyId ProxyId = 17;
+		public static readonly ProxyId ProxyId = 16;
 		ProxyId IProxy.ProxyId
 		{
 			[System.Runtime.CompilerServices.MethodImplAttribute(256)]
