+++ MemoryAllocator/_scripts.generated/InterfaceProxyGenerator/IEntityDestroySubscriberProxy.generated.cs
@@ -7,7 +7,7 @@ namespace Sapientia.TypeIndexer
 {
 	public unsafe struct IEntityDestroySubscriberProxy : IProxy
 	{
-		public static readonly ProxyId ProxyId = 22;
+		public static readonly ProxyId ProxyId = 21;
 		ProxyId IProxy.ProxyId
 		{
 			[System.Runtime.CompilerServices.MethodImplAttribute(256)]
