+++ MemoryAllocator/_scripts.generated/InterfaceProxyGenerator/IInterfaceProxyTypeProxy.generated.cs
@@ -7,7 +7,7 @@ namespace Sapientia.TypeIndexer
 {
 	public unsafe struct IInterfaceProxyTypeProxy : IProxy
 	{
-		public static readonly ProxyId ProxyId = 12;
+		public static readonly ProxyId ProxyId = 11;
 		ProxyId IProxy.ProxyId
 		{
 			[System.Runtime.CompilerServices.MethodImplAttribute(256)]
