+++ MemoryAllocator/_scripts.generated/InterfaceProxyGenerator/IKillSubscriberProxy.generated.cs
@@ -7,7 +7,7 @@ namespace Sapientia.TypeIndexer
 {
 	public unsafe struct IKillSubscriberProxy : IProxy
 	{
-		public static readonly ProxyId ProxyId = 21;
+		public static readonly ProxyId ProxyId = 20;
 		ProxyId IProxy.ProxyId
 		{
 			[System.Runtime.CompilerServices.MethodImplAttribute(256)]
