+++ MemoryAllocator/_scripts.generated/InterfaceProxyGenerator/IWorldStatePartProxy.generated.cs
@@ -7,7 +7,7 @@ namespace Sapientia.TypeIndexer
 {
 	public unsafe struct IWorldStatePartProxy : IProxy
 	{
-		public static readonly ProxyId ProxyId = 15;
+		public static readonly ProxyId ProxyId = 14;
 		ProxyId IProxy.ProxyId
 		{
 			[System.Runtime.CompilerServices.MethodImplAttribute(256)]
