+++ MemoryAllocator/_scripts.generated/InterfaceProxyGenerator/IWorldSystemProxy.generated.cs
@@ -7,7 +7,7 @@ namespace Sapientia.TypeIndexer
 {
 	public unsafe struct IWorldSystemProxy : IProxy
 	{
-		public static readonly ProxyId ProxyId = 16;
+		public static readonly ProxyId ProxyId = 15;
 		ProxyId IProxy.ProxyId
 		{
 			[System.Runtime.CompilerServices.MethodImplAttribute(256)]
