+++ MemoryAllocator/_scripts.generated/InterfaceProxyGenerator/IWorldUnmanagedLocalStatePartProxy.generated.cs
@@ -7,7 +7,7 @@ namespace Sapientia.TypeIndexer
 {
 	public unsafe struct IWorldUnmanagedLocalStatePartProxy : IProxy
 	{
-		public static readonly ProxyId ProxyId = 23;
+		public static readonly ProxyId ProxyId = 22;
 		ProxyId IProxy.ProxyId
 		{
 			[System.Runtime.CompilerServices.MethodImplAttribute(256)]
